chore: refactor events and home page
